+++ integration-test/integration/diagnostics_test.clj
@@ -11,36 +11,75 @@
   (lsp/start-process!)
   (lsp/request! (fixture/initialize-request))
   (lsp/notify! (fixture/initialized-notification))
+  (lsp/notify! (fixture/did-open-notification "diagnostics/unused_public_var.clj"))
 
   (testing "When a public var is unused"
-    (lsp/notify! (fixture/did-open-notification "diagnostics/unused_public_var.clj"))
     (h/assert-submap
       {:diagnostics
-       [{:range {:start {:line 2, :character 6}
-                 :end {:line 2, :character 9}}
+       [{:range {:start {:line 2 :character 6}
+                 :end {:line 2 :character 9}}
          :severity 3
          :code "unused-public-var"
          :source "clojure-lsp"
          :message "Unused public var 'diagnostics.unused-public-var/foo'"
          :tags [1]}
-        {:range {:start {:line 5, :character 5}
-                 :end {:line 5, :character 8}}
+        {:range {:start {:line 5 :character 5}
+                 :end {:line 5 :character 8}}
          :severity 3
          :code "unused-public-var"
          :source "clojure-lsp"
          :message "Unused public var 'diagnostics.unused-public-var/bar'"
          :tags [1]}]}
-      (lsp/await-notification :textDocument/publishDiagnostics)))
+      (lsp/await-notification :textDocument/publishDiagnostics))))
 
-  (testing "When a private var is unused"
-    (lsp/notify! (fixture/did-open-notification "diagnostics/kondo.clj"))
+(deftest report-duplicates-enabled
+  (lsp/start-process!)
+  (lsp/request! (fixture/initialize-request))
+  (lsp/notify! (fixture/initialized-notification))
+  (lsp/notify! (fixture/did-open-notification "diagnostics/kondo.clj"))
+
+  (testing "when report-duplicates is enabled by default"
     (h/assert-submap
       {:diagnostics
-       [{:range {:start {:line 2 :character 16}
-                 :end {:line 2 :character 19}}
+       [{:range {:start {:line 2 :character 16} :end {:line 2 :character 19}}
          :severity 2
          :code "unused-private-var"
          :source "clj-kondo"
          :message "Unused private var diagnostics.kondo/foo"
-         :tags [1]}]}
+         :tags [1]}
+        {:range {:start {:line 5 :character 0} :end {:line 5 :character 3}}
+         :severity 1
+         :code "unresolved-symbol"
+         :source "clj-kondo"
+         :message "Unresolved symbol: bar"
+         :tags []}
+        {:range {:start {:line 7 :character 0} :end {:line 7 :character 3}}
+         :severity 1
+         :code "unresolved-symbol"
+         :source "clj-kondo"
+         :message "Unresolved symbol: bar"
+         :tags []}]}
+      (lsp/await-notification :textDocument/publishDiagnostics))))
+
+(deftest report-duplicates-disabled
+  (lsp/start-process!)
+  (lsp/request! (fixture/initialize-request {:linters {:clj-kondo {:report-duplicates false}}}))
+  (lsp/notify! (fixture/initialized-notification))
+  (lsp/notify! (fixture/did-open-notification "diagnostics/kondo.clj"))
+
+  (testing "when report-duplicates is disabled manually"
+    (h/assert-submap
+      {:diagnostics
+       [{:range {:start {:line 2 :character 16} :end {:line 2 :character 19}}
+         :severity 2
+         :code "unused-private-var"
+         :source "clj-kondo"
+         :message "Unused private var diagnostics.kondo/foo"
+         :tags [1]}
+        {:range {:start {:line 5 :character 0} :end {:line 5 :character 3}}
+         :severity 1
+         :code "unresolved-symbol"
+         :source "clj-kondo"
+         :message "Unresolved symbol: bar"
+         :tags []}]}
       (lsp/await-notification :textDocument/publishDiagnostics))))