+++ integration-test/sample-test/src/diagnostics/kondo.clj
@@ -2,3 +2,7 @@
 
 (defn ^:private foo []
   (+ 1 2))
+
+bar
+
+bar